--- air-quality-ui/system_tests/ui/aqi_snapshot.spec.ts
@@ -1,8 +1,0 @@
-import { expect, test } from '../utils/fixtures'
-
-test('AQI snapshot assertion', async ({ vairifyCityPage }) => {
-  await vairifyCityPage.setupCityPageGraph()
-  await expect(vairifyCityPage.textFinder('Rio de Janeiro')).toBeVisible()
-  const chartShot = await vairifyCityPage.captureChartScreenshot()
-  expect(chartShot).toMatchSnapshot('rio-aqi-graph.png')
-})
